// frontend/src/typechain-v2/ISemaphore.ts
export declare namespace ISemaphore {
  export type SemaphoreProofStruct = {
    merkleTreeDepth: BigNumberish;
    merkleTreeRoot: BigNumberish;
    nullifier: BigNumberish;
    message: BigNumberish;
    scope: BigNumberish;
    points: BigNumberish[];
  };

  export type SemaphoreProofStructOutput = [
    merkleTreeDepth: bigint,
    merkleTreeRoot: bigint,
    nullifier: bigint,
    message: bigint,
    scope: bigint,
    points: bigint[]
  ] & {
    merkleTreeDepth: bigint;
    merkleTreeRoot: bigint;
    nullifier: bigint;
    message: bigint;
    scope: bigint;
    points: bigint[];
  };
}