// frontend/src/typechain-v2/frontend/abis/ISemaphoreGroups.ts
  export type InputTuple = [
    groupId: BigNumberish,
    oldAdmin: AddressLike,
    newAdmin: AddressLike
  ];
  export type OutputTuple = [
    groupId: bigint,
    oldAdmin: string,
    newAdmin: string
  ];
  export interface OutputObject {
    groupId: bigint;
    oldAdmin: string;
    newAdmin: string;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
// ...
export namespace MemberAddedEvent {
  export type InputTuple = [
    groupId: BigNumberish,
    index: BigNumberish,
    identityCommitment: BigNumberish,
    merkleTreeRoot: BigNumberish
  ];
  export type OutputTuple = [
    groupId: bigint,
    index: bigint,
    identityCommitment: bigint,
    merkleTreeRoot: bigint
  ];
  export interface OutputObject {
    groupId: bigint;
    index: bigint;
    identityCommitment: bigint;
    merkleTreeRoot: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export namespace MemberRemovedEvent {
  export type InputTuple = [
    groupId: BigNumberish,
    index: BigNumberish,
    identityCommitment: BigNumberish,
    merkleTreeRoot: BigNumberish
  ];
  export type OutputTuple = [
    groupId: bigint,
    index: bigint,
    identityCommitment: bigint,
    merkleTreeRoot: bigint
  ];
  export interface OutputObject {
    groupId: bigint;
    index: bigint;
    identityCommitment: bigint;
    merkleTreeRoot: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export namespace MemberUpdatedEvent {
  export type InputTuple = [
    groupId: BigNumberish,
    index: BigNumberish,
    identityCommitment: BigNumberish,
    newIdentityCommitment: BigNumberish,
    merkleTreeRoot: BigNumberish
  ];
  export type OutputTuple = [
    groupId: bigint,
    index: bigint,
    identityCommitment: bigint,
    newIdentityCommitment: bigint,
    merkleTreeRoot: bigint
  ];
  export interface OutputObject {
    groupId: bigint;
    index: bigint;
    identityCommitment: bigint;
    newIdentityCommitment: bigint;
    merkleTreeRoot: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export namespace MembersAddedEvent {
  export type InputTuple = [
    groupId: BigNumberish,
    startIndex: BigNumberish,
    identityCommitments: BigNumberish[],
    merkleTreeRoot: BigNumberish
  ];
  export type OutputTuple = [
    groupId: bigint,
    startIndex: bigint,
    identityCommitments: bigint[],
    merkleTreeRoot: bigint
  ];
  export interface OutputObject {
    groupId: bigint;
    startIndex: bigint;
    identityCommitments: bigint[];
    merkleTreeRoot: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}